refactor(tests): split txt checker suite by scenario
Replace the monolithic TXT checker tests with required, optional, and error-focused modules, add shared TXT test support helpers, and strengthen negative-path assertions with stable message checks while preserving existing TXT test coverage.
